separated site from config [wip]
